scripts/get_data/import_zones.py
import psycopg2, csv, os

host = 'postgres'
dbname = os.getenv('POSTGRES_DB')
user = os.getenv('POSTGRES_USER')
password = os.getenv('POSTGRES_PASSWORD')
conn = psycopg2.connect(f"host={host} dbname={dbname} user={user} password={password}")
cur = conn.cursor()

with open('zones.csv', 'r') as f:
    reader = csv.reader(f)
    next(reader) # Skip the header row.
    for row in reader:
        cur.execute(
        "INSERT INTO zone (zone_id,zone_name,country,matchup) VALUES (%s, %s, %s, %s)",
        row
    )
conn.commit()